requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.6.3
cachelib==0.10.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
distro==1.8.0
dnspython==2.3.0
eventlet==0.33.3
Flask==2.2.5
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
frozenlist==1.3.3
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==1.2.2
packaging==23.2
pdfminer.six==20221105
ProWritingAidSDK @ git+https://github.com/prowriting/prowritingaid.python.git@fd6fc238778b7b5ba4bb56835d3ed73594f3bcb7
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-engineio==4.8.0
python-socketio==5.10.0
redis==5.0.1
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wsproto==1.2.0
yarl==1.9.2
